Rename `PlatformTitlebar` to `TitleBar`
