+++ crates/storybook/src/story_selector.rs
@@ -29,10 +29,10 @@ pub enum ComponentStory {
     ListHeader,
     ListItem,
     OverflowScroll,
-    PlatformTitlebar,
     Scroll,
     Tab,
     TabBar,
+    TitleBar,
     ToggleButton,
     Text,
     ViewportUnits,
@@ -61,11 +61,11 @@ impl ComponentStory {
             Self::ListHeader => cx.new_view(|_| ui::ListHeaderStory).into(),
             Self::ListItem => cx.new_view(|_| ui::ListItemStory).into(),
             Self::OverflowScroll => cx.new_view(|_| crate::stories::OverflowScrollStory).into(),
-            Self::PlatformTitlebar => cx.new_view(|_| ui::PlatformTitlebarStory).into(),
             Self::Scroll => ScrollStory::view(cx).into(),
             Self::Text => TextStory::view(cx).into(),
             Self::Tab => cx.new_view(|_| ui::TabStory).into(),
             Self::TabBar => cx.new_view(|_| ui::TabBarStory).into(),
+            Self::TitleBar => cx.new_view(|_| ui::TitleBarStory).into(),
             Self::ToggleButton => cx.new_view(|_| ui::ToggleButtonStory).into(),
             Self::ViewportUnits => cx.new_view(|_| crate::stories::ViewportUnitsStory).into(),
             Self::Picker => PickerStory::new(cx).into(),
