+++ crates/ui/src/components/stories.rs
@@ -10,9 +10,9 @@ mod label;
 mod list;
 mod list_header;
 mod list_item;
-mod platform_titlebar;
 mod tab;
 mod tab_bar;
+mod title_bar;
 mod toggle_button;
 
 pub use avatar::*;
@@ -27,7 +27,7 @@ pub use label::*;
 pub use list::*;
 pub use list_header::*;
 pub use list_item::*;
-pub use platform_titlebar::*;
 pub use tab::*;
 pub use tab_bar::*;
+pub use title_bar::*;
 pub use toggle_button::*;
