+++ crates/ui/src/components/title_bar.rs
@@ -1,0 +1,3 @@
+mod title_bar;
+
+pub use title_bar::*;
